INT-3570 Core 4.0 Schema Work Around
JIRA: https://jira.spring.io/browse/INT-3570

In 4.1 we removed the 'old' schemas to reduce clutter. However certain
extensions such as kafka import the 4.0 schema which is no longer on
the classpath when using Spring Integration 4.1 when there is no internet
connection.

INT-3570 states that we should restore the 4.0 core schema and mapping
to resolve this issue. In the meantime, this work around adds the schema
to the kafka jar.

It should be reverted when INT-3570 is released in Spring Integration 4.1.1.