INT-3614 Fix `extractPayload` for JMS i-c-adapter
JIRA: https://jira.spring.io/browse/INT-3614

The `extractPayload` is presented in the namespace support for `<int-jms:inbound-channel-adapter>`,
but nor `JmsInboundChannelAdapterParser`, neither `JmsDestinationPollingSource` take care of that value.

To be consist with `<int-jms:message-driven-channel-adapter>` and documentation introduce that option for
`JmsDestinationPollingSource` and populate it from the `JmsInboundChannelAdapterParser`.

**Cherry-pick to 4.1.x, 4.0.x and 3.0.x**

Note: 3.0.x will be with conflicts: there is still no `getMessageBuilderFactory()`