INT-3796: Configure SFTP Unknown Host Behavior
JIRA: https://jira.spring.io/browse/INT-3796

Previously, the `DefaultSftpSessionFactory` unconditionally accepted keys
from unknown hosts, or when a host key changed.

Permit this behavior to be configured with `allowUnknownHosts` and/or a custom
`UserInfo` properties in the factory.

`allowUknownHosts` is false by default (since 4.2) and true by default in 4.1.x for
backwards compatibility.

INT-3796: Fix Tests

INT-3796: Polishing

Conflicts:
	src/reference/asciidoc/sftp.adoc
	src/reference/asciidoc/whats-new.adoc

Fix `SftpSessionFactoryTests` according to the `allowUnknownKeys = true` in the `4.1.x`