INT-3619: Prevent `scanner` options overriding
JIRA: https://jira.spring.io/browse/INT-3619

Previously the `FileReadingMessageSource` allowed to override `filter` and `locker` properties
on the `scanner` through its own setters.
It caused confusing and unexpected behavior.

Provide refactoring to check that provided `filter` and `locker` are applied for an internal `scanner`
instance, not for the injected one.