INT-3534:`MethodInvokingMLP` don't wrap messages
JIRA: https://jira.spring.io/browse/INT-3534

Previously the `MethodInvokingMessageListProcessor` wrapped `Collection<Message<?>>`
in its `process()` method for the `delegate` to a new `Collection` unconditionally.

Since the better performance is one of our goals, it would be better do not wrap like it is
in the `ExpressionEvaluatingMessageGroupProcessor`.

Also add an `important` note to the `aggregator.adoc` about the restriction with `unmodifiableCollection`.